refactor(cli-config): Invert `clap-utils` dependency
Additionally deduplicate `normalize_to_url_if_moniker`
