+++ clap-utils/src/input_validators.rs
@@ -209,16 +209,7 @@ where
     }
 }
 
-pub fn normalize_to_url_if_moniker<T: AsRef<str>>(url_or_moniker: T) -> String {
-    match url_or_moniker.as_ref() {
-        "m" | "mainnet-beta" => "https://api.mainnet-beta.solana.com",
-        "t" | "testnet" => "https://api.testnet.solana.com",
-        "d" | "devnet" => "https://api.devnet.solana.com",
-        "l" | "localhost" => "http://localhost:8899",
-        url => url,
-    }
-    .to_string()
-}
+pub use solana_cli_config::normalize_to_url_if_moniker;
 
 pub fn is_epoch<T>(epoch: T) -> Result<(), String>
 where
