+++ clap-v3-utils/Cargo.toml
@@ -27,6 +27,7 @@ chrono = { workspace = true, features = ["default"] }
 clap = { version = "3.2.23", default-features = false, features = ["cargo", "std", "suggestions"] }
 five8 = { workspace = true }
 rpassword = { workspace = true }
+solana-cli-config = { workspace = true }
 solana-clock = { workspace = true }
 solana-cluster-type = { workspace = true }
 solana-commitment-config = { workspace = true }
