+++ cli-config/src/config_input.rs
@@ -1,7 +1,4 @@
-use {
-    crate::Config, solana_clap_utils::input_validators::normalize_to_url_if_moniker,
-    solana_commitment_config::CommitmentConfig, std::str::FromStr,
-};
+use {crate::Config, solana_commitment_config::CommitmentConfig, std::str::FromStr};
 
 pub enum SettingType {
     Explicit,
@@ -123,4 +120,15 @@ impl Default for ConfigInput {
             commitment: CommitmentConfig::confirmed(),
         }
     }
+}
+
+pub fn normalize_to_url_if_moniker<T: AsRef<str>>(url_or_moniker: T) -> String {
+    match url_or_moniker.as_ref() {
+        "m" | "mainnet-beta" => "https://api.mainnet-beta.solana.com",
+        "t" | "testnet" => "https://api.testnet.solana.com",
+        "d" | "devnet" => "https://api.devnet.solana.com",
+        "l" | "localhost" => "http://localhost:8899",
+        url => url,
+    }
+    .to_string()
 }
